Naming things: Refactor package and docs layout, by adding `source`
