+++ docs/backlog.md
@@ -1,10 +1,7 @@
 # Backlog
 
 ## Iteration +1
-- Shell '25: Introduce _daily_ and _weekly_ operation modes / interfaces
-- Naming things: `source` vs. `sink` vs. Slack CLI domains
-- Project layout and docs: Bring up to speed
-- Docs: Generic page about `--when` option
+- Caveats / Philosophy
 - Shell/Daily: Interleave CI reports
 - UI/Slack: Improve zapping
 - Aika: Fix bugs and refactor code
@@ -14,6 +11,7 @@
 - UI/Slack: chat_scheduleMessage
 - Options: Program currently understands `--slack-token=`, but lacks `--github-token=`
 - UI/Console: Spice up Markdown output using `rich` and friends
+- `rapporto gh activity` needs a section about "Top issues"
 
 ## Iteration +2
 - GitHub/Actions: Currently lacks parameter `--when`
@@ -26,6 +24,7 @@
 - GitHub: Search items with people assignments
 - GitHub/API: On errors, the JSON response includes the reason as an
   error message. However, it isn't displayed, yet.
+- GitHub: Report about stale issues
 - Options: Make `limit=100` configurable? Is paging needed?
 - `github-backup` can do "Exceeded rate limit of 5000 requests;
   waiting 77 seconds to reset". Do we also need it?
@@ -81,3 +80,7 @@
 - Options: Parse time intervals using `aika`
 - Slack: Support for `mrkdwn` output
 - GitHub: Process `created` + `updated`, not just `created`
+- Shell '25: Introduce _daily_ and _weekly_ operation modes / interfaces
+- Docs: Generic page about `--when` option
+- Naming things: `source` vs. `sink` vs. Slack CLI domains
+- Project layout and docs: Bring up to speed
