+++ src/rapporto/source/slack/core.py
